+++ Backpack/Theme/Classes/BPKHongKongTheme.m
@@ -46,6 +46,10 @@
     return self.primaryColor;
 }
 
+- (UIColor *)buttonLinkContentColor {
+    return self.primaryColor;
+}
+
 - (UIColor *)buttonPrimaryContentColor {
     return [UIColor colorWithRed:255.0f / 255.0f green:163.0f / 255.0f blue:203.0f / 255.0f alpha:1.0f];
 }